Finstock/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
bcrypt==3.1.7
bidict==0.23.1
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==2.8
distlib==0.3.8
Django==4.2.14
django-cors-headers==4.4.0
Fabric3==1.14.post1
filelock==3.13.4
Flask==2.1.0
Flask-Cors==4.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mypy==1.11.1
mypy-extensions==1.0.0
paramiko==2.12.0
pbr==6.0.0
pep8==1.7.1
platformdirs==4.2.0
pycodestyle==2.11.1
pycparser==2.22
PyNaCl==1.3.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
spark-parser==1.8.9
SQLAlchemy==2.0.29
sqlparse==0.5.0
stevedore==5.2.0
tomli==2.0.1
typing_extensions==4.11.0
uncompyle2==2.0.0
uncompyle6==3.9.0
urllib3==2.2.1
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
Werkzeug==2.1.0
wsproto==1.2.0
xdis==6.0.5
yarl==1.9.4
zipp==3.18.1
django-environ
daphne
djangorestframework
django-filter
reportlab
drf-yasg
mysqlclient
